requirements.txt
fastapi==0.78.0
pydantic==1.9.0
pymongo==4.1.1
python_dateutil==2.8.2
requests==2.27.1
